ubuntu-dev: move to gcc-arm armel package
Starting from ubuntu 22.04, 'gcc-arm-linux-gnueabihf' enforces hard
floating point making the linux kernel build to fail. The reason is that
the build system won't be able to compile for '-march=armv7-a' and it will
fallback to 'armv5t' giving all sort of errors due to unsupported
instructions:

"
/tmp/ccMeKsEN.s:2201: Error: selected processor does not support `cpsid
i' in ARM mode
/tmp/ccMeKsEN.s:3069: Error: selected processor does not support `cpsid
i' in ARM mode
/tmp/ccMeKsEN.s:3279: Error: selected processor does not support `cpsie
i' in ARM mode
/tmp/ccMeKsEN.s:3446: Error: selected processor does not support `dsb '
in ARM mode
/tmp/ccMeKsEN.s:3910: Error: selected processor does not support `cpsie
i' in ARM mode
/tmp/ccMeKsEN.s:4286: Error: architectural extension `mp' is not allowed
for the current base architecture
/tmp/ccMeKsEN.s:4287: Error: selected processor does not support `pldw
[r4]' in ARM mode
/tmp/ccMeKsEN.s:4290: Error: selected processor does not support `pld
[r4]' in ARM mode
/tmp/ccMeKsEN.s:4304: Error: selected processor does not support `ldrex
r3,[r4]' in ARM mode
/tmp/ccMeKsEN.s:4306: Error: selected processor does not support `strex
r1,r2,[r4]' in ARM mode
"

Using the gcc-arm armel package solves the above issue...

I'm not sure what will be the behavior in future Ubuntu release but as
long as the armel package works, I do not intend to change back this
package.

Signed-off-by: Nuno Sá <noname.nuno@gmail.com>